Rename all .h files to .hpp and update includes
