+++ dx2/empty.cxx
@@ -1,2 +1,2 @@
 #include <Eigen/Dense>
-#include <dx2/dx2.h>
+#include <dx2/dx2.hpp>
