+++ include/dx2/beam.hpp
