+++ include/dx2/goniometer.hpp
